use lu and typecast to avoid warning for sizeof() between 32 and 64 bit hosts

+++ lib/hcrypto/tomsfastmath/src/misc/fp_ident.c
@@ -18,8 +18,8 @@ const char *fp_ident(void)
 "TomsFastMath (%s)\n"
 "\n"
 "Sizeofs\n"
-"\tfp_digit = %u\n"
-"\tfp_word  = %u\n"
+"\tfp_digit = %lu\n"
+"\tfp_word  = %lu\n"
 "\n"
 "FP_MAX_SIZE = %u\n"
 "\n"
@@ -70,7 +70,7 @@ const char *fp_ident(void)
 #ifdef TFM_HUGE
 " TFM_HUGE "
 #endif
-"\n", __DATE__, sizeof(fp_digit), sizeof(fp_word), FP_MAX_SIZE);
+"\n", __DATE__, (unsigned long)sizeof(fp_digit), (unsigned long)sizeof(fp_word), FP_MAX_SIZE);
 
    if (sizeof(fp_digit) == sizeof(fp_word)) {
       strncat(buf, "WARNING: sizeof(fp_digit) == sizeof(fp_word), this build is likely to not work properly.\n",
